feat: allow --reporter to be relative to CWD
Previously, the --reporter path must have been either absolute or in
node_modules or relative to lib/compiler.js.

This change ensures that a path passed to --reporter is resolved relative
to process.cwd().

e.g. you can now run this inside of a project directory to use a custom
reporter:

    webpack --reporter ./MyCustomReporter.js